Initalization of GitHub Repository
Setting up GitHub remote repository and adding code that serves as a work in progress for the moment.  This is needed at this point in the process to make sure the .csv file can be accessed from GitHub to make the necessary Mata object(s) for browsing, searching, and selecting data sets of interest.