Redesigned Ipeds object
Arrays seem to be acting as wrappers for a pointer and can't be saved directly.  New approach is to store binary matrix on local disk and read that into Mata directly.  From there pointers will be created to the different search vectors to make searching/getting a bit more efficient.